Active link, auth manageable webinar

+++ app/Http/Controllers/WebinarsController.php
@@ -2,8 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\User;
 use App\Webinar;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class WebinarsController extends Controller
 {
@@ -71,26 +73,30 @@ class WebinarsController extends Controller
     {
         $this->authorize('maintainer control');
         $req = $this->validation($request);
-        $ins = Webinar::create($req);
+        $req['creator_id'] = Auth::user()->id;
+        Webinar::create($req);
         return back()->with('success', trans('You have successfully creating new webinar!'));
     }
 
     public function manage()
     {
         $this->authorize('maintainer control');
-        $webinars = Webinar::paginate(2);
+        $webinars = Webinar::paginate(4);
         return view('admin.manage_webinars', [ 'webinars' => $webinars ]);
     }
 
     public function edit(Webinar $webinar)
     {
         $this->authorize('maintainer control');
+        $this->authorize('update-webinar', $webinar);
+        $user = Auth::user();
         return view('admin.edit_webinar', [ 'webinar' => $webinar ]);
     }
 
     public function update(Request $request, Webinar $webinar)
     {
         $this->authorize('maintainer control');
+        $this->authorize('update-webinar', $webinar);
         $req = $this->validation($request);
         $webinar->update($req);
         return back()->with('success', trans('You have successfully updated ').$webinar->nama);
@@ -99,6 +105,7 @@ class WebinarsController extends Controller
     public function destroy(Webinar $webinar)
     {
         $this->authorize('maintainer control');
+        $this->authorize('update-webinar', $webinar);
         $webinar->delete();
         return back()->with('success', trans('You have successfully deleted ').$webinar->nama);
     }
